+++ tencentcloud/ckafka/v20190819/models.go
@@ -7407,6 +7407,9 @@ type MySQLParam struct {
 
 	// 如果该值为true，且MySQL中"binlog_rows_query_log_events"配置项的值为"ON"，则流入到topic的数据包含原SQL语句；若该值为false，流入到topic的数据不包含原SQL语句
 	IncludeQuery *bool `json:"IncludeQuery,omitempty" name:"IncludeQuery"`
+
+	// 如果该值为 true，则消息中会携带消息结构体对应的schema，如果该值为false则不会携带
+	RecordWithSchema *bool `json:"RecordWithSchema,omitempty" name:"RecordWithSchema"`
 }
 
 type OperateResponseData struct {
@@ -7529,6 +7532,21 @@ type PostgreSQLParam struct {
 
 	// 复制存量信息(never增量, initial全量)，默认为initial
 	SnapshotMode *string `json:"SnapshotMode,omitempty" name:"SnapshotMode"`
+
+	// 上游数据格式(JSON/Debezium), 当数据库同步模式为默认字段匹配时,必填
+	DataFormat *string `json:"DataFormat,omitempty" name:"DataFormat"`
+
+	// "INSERT" 表示使用 Insert 模式插入，"UPSERT" 表示使用 Upsert 模式插入
+	DataTargetInsertMode *string `json:"DataTargetInsertMode,omitempty" name:"DataTargetInsertMode"`
+
+	// 当 "DataInsertMode"="UPSERT" 时，传入当前 upsert 时依赖的主键
+	DataTargetPrimaryKeyField *string `json:"DataTargetPrimaryKeyField,omitempty" name:"DataTargetPrimaryKeyField"`
+
+	// 表与消息间的映射关系
+	DataTargetRecordMapping []*RecordMapping `json:"DataTargetRecordMapping,omitempty" name:"DataTargetRecordMapping"`
+
+	// 是否抛弃解析失败的消息，默认为true
+	DropInvalidMessage *bool `json:"DropInvalidMessage,omitempty" name:"DropInvalidMessage"`
 }
 
 type Price struct {